backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Mamadou Ahmadou ka",email = "mahmadou.ka@univ-thies.sn"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pyyaml (>=6.0.3,<7.0.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "sqlmodel (>=0.0.31,<0.0.32)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "langchain (>=1.2.3,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-huggingface (>=1.2.0,<2.0.0)",
    "langchain-qdrant (>=1.1.0,<2.0.0)",
    "pypdf (>=6.6.0,<7.0.0)",
    "sentence-transformers (>=5.2.0,<6.0.0)",
    "langchain-groq (>=1.1.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "bcrypt (==4.0.1)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
